python/iotsitewise/model_limit_test/requirements.txt
aiobotocore==2.17.0
anytree==2.12.1
asyncstdlib==3.13.0
dynamodb-json==1.4.2
python-lambda==11.8.0
